Additional work:  - moves all models from root directory to js/angular
