+++ src/js/angular/models/connectors/connector-command.js
