+++ src/js/angular/models/jdbc/jdbc-column.js
