+++ src/js/angular/models/jdbc/jdbc-configuration-error.js
