+++ src/js/angular/models/jdbc/jdbc-configuration-info.js
