+++ src/js/angular/models/ontop/ontop-connection-information.js
