+++ src/js/angular/models/ontop/ontop-file-info.js
