+++ src/js/angular/models/ontop/ontop-file-type.js
