+++ src/js/angular/models/ontotext-yasgui/before-update-query-result.js
