+++ src/js/angular/models/ontotext-yasgui/count-query-request-event.js
