+++ src/js/angular/models/ontotext-yasgui/count-query-response-event.js
