+++ src/js/angular/models/ontotext-yasgui/download-as-event.js
