+++ src/js/angular/models/ontotext-yasgui/event-data-type.js
