+++ src/js/angular/models/ontotext-yasgui/event-data.js
