+++ src/js/angular/models/ontotext-yasgui/notification-message-event.js
