+++ src/js/angular/models/ontotext-yasgui/notification-message-type.js
