+++ src/js/angular/models/ontotext-yasgui/query-executed-event.js
