+++ src/js/angular/models/ontotext-yasgui/query-mode.js
