+++ src/js/angular/models/ontotext-yasgui/query-request-event.js
