+++ src/js/angular/models/ontotext-yasgui/query-type.js
