+++ src/js/angular/models/ontotext-yasgui/rendering-mode.js
