+++ src/js/angular/models/ontotext-yasgui/yasqe-mode.js
