+++ src/js/angular/models/ontotext-yasgui/yasr-plugin-name.js
