+++ src/js/angular/models/sparql-template/sparql-template-error.js
