+++ src/js/angular/models/sparql-template/sparql-template-info.js
