+++ src/js/angular/models/sparql/sparql-query-error-info.js
